fix(engine): move SecuredRequestHandler out of common/ (dev/app mirror has no exceptions/ tree) (issue #602)
