+++ source/spec/lib/server/handlers/api/ApiConfigHandler_spec.js
@@ -1,8 +1,8 @@
-import { SecuredRequestHandler } from '../../../../../lib/common/server/SecuredRequestHandler.js';
 import { ClientNotFound } from '../../../../../lib/exceptions/registry/ClientNotFound.js';
 import { ResourceNotFound } from '../../../../../lib/exceptions/registry/ResourceNotFound.js';
 import { NamespaceMap } from '../../../../../lib/registry/NamespaceMap.js';
 import { ApiConfigHandler } from '../../../../../lib/server/handlers/api/ApiConfigHandler.js';
+import { SecuredRequestHandler } from '../../../../../lib/server/SecuredRequestHandler.js';
 import { Application } from '../../../../../lib/services/Application.js';
 import { ResourceEnqueuer } from '../../../../../lib/utils/ResourceEnqueuer.js';
